requirements.txt
pytorch=1.6.0
torchvision=0.7.0
numpy=1.19.1
scipy=1.5.2
scikit-learn=0.23.2
matplotlib=3.3.2
pillow=7.2.0
h5py=2.10.0
opencv-python=4.4.0.44
tqdm=4.50.2
pyyaml=5.3.1
protobuf=3.17.3
typing-extensions=3.10.0.0
numexpr=2.7.1
joblib=0.17.0
matplotlib-base=3.3.2
networkx=2.5
tensorflow=2.5.0
keras=2.4.3
pywavelets=1.1.1
toolz=0.11.1
dataclasses=0.8
einops=0.3.0
cycler=0.10.0
kiwisolver=1.2.0
astropy=4.0.1.post1
tensorflow-data-validation=1.3.0
google-auth=1.32.1
pydicom=2.1.1
simpleitk=2.0.2
protobuf=3.17.3
torch-radon=1.0.0
tensorflow-estimator=2.5.0
tensorflow-metadata=1.1.0
tensorflow-probability=0.12.1
ptflops=0.6.4
kornia=0.4.0
pyfftw=0.12.0
healpy=1.14.0
markdown=3.3.4
tensorflow-hub=0.12.0